requirements.txt
# Core libraries
numpy==1.23.5
opencv-python==4.10.0.84
pillow==10.4.0
matplotlib==3.10.3

# Flask server
flask==2.3.3

# SocketIO for simulator
python-socketio==4.6.1
python-engineio==3.13.2
eventlet==0.40.3

# Deep learning
tensorflow==2.10.0
keras==2.11.0

# Optional: for data manipulation
pandas==2.1.1
scikit-learn==1.3.0
